Extract CLI definition, input handling and error management
