+++ src/main.rs
@@ -1,454 +1,21 @@
-use atty::Stream;
 use biscuit_auth::{
-    builder::{BiscuitBuilder, BlockBuilder, Policy},
+    builder::Policy,
     error::{FailedCheck, Logic, MatchedPolicy, RunLimit, Token},
-    parser::{parse_block_source, parse_source},
-    Authorizer, Biscuit, UnverifiedBiscuit, {KeyPair, PrivateKey, PublicKey},
+    Biscuit, {KeyPair, PrivateKey},
 };
 use clap::Parser;
-use std::env;
 use std::error::Error;
-use std::fmt;
-use std::fmt::Display;
-use std::fs;
 use std::io::Write;
-use std::io::{self, Read};
+use std::io::{self};
 use std::path::PathBuf;
-use std::process::Command;
 
-/// biscuit creation and inspection CLI. Run `biscuit --help` to see what's available.
-#[derive(Parser)]
-#[clap(version = "1.0", author = "Clément D. <clement@delafargue.name>")]
-struct Opts {
-    // /// Sets a custom config file. Could have been an Option<T> with no default too
-    // #[clap(short, long, default_value = "default.conf")]
-    // config: String,
-    // /// Some input. Because this isn't an Option<T> it's required to be used
-    // input: String,
-    // /// A level of verbosity, and can be used multiple times
-    // #[clap(short, long, parse(from_occurrences))]
-    // verbose: i32,
-    #[clap(subcommand)]
-    subcmd: SubCommand,
-}
+mod cli;
+mod errors;
+mod input;
 
-#[derive(Parser)]
-enum SubCommand {
-    #[clap(name = "keypair")]
-    KeyPairCmd(KeyPairCmd),
-    #[clap()]
-    Inspect(Inspect),
-    #[clap()]
-    Generate(Generate),
-    #[clap()]
-    Attenuate(Attenuate),
-}
-
-/// Create and manipulate key pairs
-#[derive(Parser)]
-struct KeyPairCmd {
-    /// Generate the keypair from the given private key. If omitted, a random keypair will be generated
-    #[clap(long, conflicts_with("from-private-key-file"))]
-    from_private_key: Option<String>,
-    /// Generate the keypair from a private key stored in the given file (or use `-` to read it from stdin). If omitted, a random keypair will be generated
-    #[clap(long, parse(from_os_str))]
-    from_private_key_file: Option<PathBuf>,
-    /// Read the private key raw bytes directly, with no hex decoding
-    #[clap(long, requires("from-private-key-file"))]
-    from_raw_private_key: bool,
-    /// Only output the public part of the key pair
-    #[clap(long, conflicts_with("only-private-key"))]
-    only_public_key: bool,
-    /// Output the public key raw bytes directly, with no hex encoding
-    #[clap(long, requires("only-public-key"))]
-    raw_public_key_output: bool,
-    /// Only output the public part of the key pair
-    #[clap(long, conflicts_with("only-public-key"))]
-    only_private_key: bool,
-    /// Output the private key raw bytes directly, with no hex encoding
-    #[clap(long, requires("only-private-key"))]
-    raw_private_key_output: bool,
-}
-
-/// Generate a biscuit from a private key and an authority block
-#[derive(Parser)]
-struct Generate {
-    /// Read the authority block from the given file (or use `-` to read from stdin). If omitted, an interactive $EDITOR will be opened.
-    #[clap(parse(from_os_str))]
-    authority_file: Option<PathBuf>,
-    /// Output the biscuit raw bytes directly, with no base64 encoding
-    #[clap(long)]
-    raw: bool,
-    /// The private key used to sign the token
-    #[clap(long, required_unless_present("private-key-file"))]
-    private_key: Option<String>,
-    /// The private key used to sign the token
-    #[clap(
-        long,
-        parse(from_os_str),
-        required_unless_present("private-key"),
-        conflicts_with = "private-key"
-    )]
-    private_key_file: Option<PathBuf>,
-    /// Read the private key raw bytes directly (only available when reading the private key from a file)
-    #[clap(long, conflicts_with = "private-key", requires = "private-key-file")]
-    raw_private_key: bool,
-    /// The optional context string attached to the authority block
-    #[clap(long)]
-    context: Option<String>,
-}
-
-/// Attenuate an existing biscuit by adding a new block
-#[derive(Parser)]
-struct Attenuate {
-    /// Read the biscuit from the given file (or use `-` to read from stdin)
-    #[clap(parse(from_os_str))]
-    biscuit_file: PathBuf,
-    /// Read the biscuit raw bytes directly, with no base64 parsing
-    #[clap(long)]
-    raw_input: bool,
-    /// Output the biscuit raw bytes directly, with no base64 encoding
-    #[clap(long)]
-    raw_output: bool,
-    /// The block to append to the token. If `--block` and `--block-file` are omitted, an interactive $EDITOR will be opened.
-    #[clap(long)]
-    block: Option<String>,
-    /// The block to append to the token. If `--block` and `--block-file` are omitted, an interactive $EDITOR will be opened.
-    #[clap(long, parse(from_os_str), conflicts_with = "block")]
-    block_file: Option<PathBuf>,
-    /// The optional context string attached to the new block
-    #[clap(long)]
-    context: Option<String>,
-}
-
-/// Inspect a biscuit and optionally check its public key
-#[derive(Parser)]
-struct Inspect {
-    /// Read the biscuit from the given file (or use `-` to read from stdin)
-    #[clap(parse(from_os_str))]
-    biscuit_file: PathBuf,
-    /// Read the biscuit raw bytes directly, with no base64 parsing
-    #[clap(long)]
-    raw_input: bool,
-    /// Check the biscuit public key
-    #[clap(long, conflicts_with("public-key-file"))]
-    public_key: Option<String>,
-    /// Check the biscuit public key
-    #[clap(long, conflicts_with("public-key"), parse(from_os_str))]
-    public_key_file: Option<PathBuf>,
-    /// Read the public key raw bytes directly
-    #[clap(long, requires("public-key-file"), conflicts_with("public-key"))]
-    raw_public_key: bool,
-    /// Open $EDITOR to provide a authorizer.
-    #[clap(
-        long,
-        conflicts_with("verify-with"),
-        conflicts_with("verify-with-file")
-    )]
-    verify_interactive: bool,
-    /// Verify the biscuit with the provided authorizer.
-    #[clap(
-        long,
-        conflicts_with("verify-with"),
-        conflicts_with("verify-interactive")
-    )]
-    verify_with_file: Option<PathBuf>,
-    /// Verify the biscuit with the provided authorizer
-    #[clap(
-        long,
-        conflicts_with("verify-with-file"),
-        conflicts_with("verify-interactive")
-    )]
-    verify_with: Option<String>,
-}
-
-pub enum BiscuitFormat {
-    RawBiscuit,
-    Base64Biscuit,
-}
-
-pub enum KeyFormat {
-    RawBytes,
-    HexKey,
-}
-
-// `Base64String` is never constructed, but is still handled in
-// `cli.rs`
-#[allow(dead_code)]
-pub enum BiscuitBytes {
-    FromStdin(BiscuitFormat),
-    FromFile(BiscuitFormat, PathBuf),
-    Base64String(String),
-}
-
-// `FromStdin` is never constructed, but is still handled in
-// `cli.rs`
-#[allow(dead_code)]
-pub enum KeyBytes {
-    FromStdin(KeyFormat),
-    FromFile(KeyFormat, PathBuf),
-    HexString(String),
-}
-
-pub enum DatalogInput {
-    FromEditor,
-    FromStdin,
-    FromFile(PathBuf),
-    DatalogString(String),
-}
-
-fn ensure_no_input_conflict(
-    datalog: &DatalogInput,
-    biscuit: &BiscuitBytes,
-) -> Result<(), Box<dyn Error>> {
-    match (datalog, biscuit) {
-        // running $EDITOR as a child process requires a working stdin. When contents from stdin has already been read, this is
-        // not the case. This could be handled by reopening stdin on /dev/tty, but it's not portable (and as such, more complicated
-        // to do in rust than just disallowing a fringe use-case)
-        (DatalogInput::FromEditor, BiscuitBytes::FromStdin(_)) => Err(E {
-            msg: "I cannot read input from both stdin and an interactive editor. Please use proper files or flags instead.".to_owned(),
-        }
-        .into()),
-        // this combination should be prevented by the clap configuration
-        (DatalogInput::FromStdin, BiscuitBytes::FromStdin(_)) => Err(E {
-            msg: "I cannot read several pieces of input from stdin at the same time. Please use proper files or flags instead.".to_owned(),
-        }
-        .into()),
-        _ => Ok(()),
-    }
-}
-
-#[derive(Debug, Clone)]
-struct E {
-    msg: String,
-}
-
-impl Display for E {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", &self.msg)
-    }
-}
-
-impl Error for E {}
-
-fn read_stdin_string(desc: &str) -> Result<String, Box<dyn Error>> {
-    if atty::is(Stream::Stdin) && atty::is(Stream::Stderr) {
-        eprintln!("Please input a {}, followed by <enter> and ^D", &desc);
-    }
-    let mut buffer = String::new();
-    io::stdin().read_to_string(&mut buffer)?;
-    Ok(buffer.trim().to_owned())
-}
-
-fn read_stdin_bytes() -> Result<Vec<u8>, Box<dyn Error>> {
-    if atty::is(Stream::Stdin) {
-        return Err(E { msg: "Can't read binary content from an interactive terminal. Please pipe the content or use a proper file.".to_owned() }.into());
-    }
-    let mut buffer = Vec::new();
-    io::stdin().read_to_end(&mut buffer).map(|_| ())?;
-    Ok(buffer)
-}
-
-fn read_editor_string() -> Result<String, Box<dyn Error>> {
-    let file = tempfile::Builder::new()
-        .suffix(".biscuit-datalog")
-        .tempfile()?;
-    let path = &file.path();
-
-    if atty::isnt(Stream::Stdin) || atty::isnt(Stream::Stdout) {
-        return Err(E {
-            msg: "Can't start an editor outside of an interactive terminal.".to_owned(),
-        }
-        .into());
-    }
-
-    let result = get_editor_command()?.arg(&path).spawn()?.wait()?;
-
-    if result.success() {
-        Ok(fs::read_to_string(&path)?)
-    } else {
-        Err(E {
-            msg: "Failed reading the datalog temporary file".to_owned(),
-        }
-        .into())
-    }
-}
-
-fn get_editor_command() -> Result<Command, Box<dyn Error>> {
-    let editor_unparsed = match env::var("EDITOR") {
-        Ok(e) => Ok(e),
-        Err(env::VarError::NotPresent) => Ok("vim".to_owned()),
-        e => e,
-    }?;
-
-    let editor_parts = shell_words::split(&editor_unparsed)?;
-    match editor_parts.split_first() {
-        Some((editor_binary, editor_args)) => {
-            let mut editor_cmd = Command::new(editor_binary);
-            editor_cmd.args(editor_args);
-            Ok(editor_cmd)
-        }
-        None => Err(E {
-            msg: "Failed to parse EDITOR environment variable".to_owned(),
-        }.into()),
-    }
-}
-
-fn read_authority_from(
-    from: &DatalogInput,
-    context: &Option<String>,
-    builder: &mut BiscuitBuilder,
-) -> Result<(), Box<dyn Error>> {
-    let string = match from {
-        DatalogInput::FromEditor => read_editor_string()?,
-        DatalogInput::FromStdin => read_stdin_string("datalog program")?,
-        DatalogInput::FromFile(f) => fs::read_to_string(&f)?,
-        DatalogInput::DatalogString(str) => str.to_owned(),
-    };
-    let result = parse_block_source(&string).map_err(|_| E {
-        msg: "parse error".into(),
-    })?;
-
-    for (fact_str, _) in result.facts {
-        builder.add_authority_fact(fact_str)?;
-    }
-    for (rule_str, _) in result.rules {
-        builder.add_authority_rule(rule_str)?;
-    }
-    for (check_str, _) in result.checks {
-        builder.add_authority_check(check_str)?;
-    }
-
-    if let Some(ctx) = context {
-        builder.set_context(ctx.to_owned());
-    }
-
-    Ok(())
-}
-
-fn read_block_from(
-    from: &DatalogInput,
-    context: &Option<String>,
-    builder: &mut BlockBuilder,
-) -> Result<(), Box<dyn Error>> {
-    let string = match from {
-        DatalogInput::FromEditor => read_editor_string()?,
-        DatalogInput::FromStdin => read_stdin_string("datalog program")?,
-        DatalogInput::FromFile(f) => fs::read_to_string(&f)?,
-        DatalogInput::DatalogString(str) => str.to_owned(),
-    };
-    let result = parse_block_source(&string).map_err(|_| E {
-        msg: "parse error".into(),
-    })?;
-
-    for (fact_str, _) in result.facts {
-        builder.add_fact(fact_str)?;
-    }
-    for (rule_str, _) in result.rules {
-        builder.add_rule(rule_str)?;
-    }
-    for (check_str, _) in result.checks {
-        builder.add_check(check_str)?;
-    }
-
-    if let Some(ctx) = context {
-        builder.set_context(ctx.to_owned());
-    }
-
-    Ok(())
-}
-
-fn read_authorizer_from(
-    from: &DatalogInput,
-    authorizer: &mut Authorizer,
-) -> Result<(), Box<dyn Error>> {
-    let string = match from {
-        DatalogInput::FromEditor => read_editor_string()?,
-        DatalogInput::FromStdin => read_stdin_string("datalog program")?,
-        DatalogInput::FromFile(f) => fs::read_to_string(&f)?,
-        DatalogInput::DatalogString(str) => str.to_owned(),
-    };
-    let result = parse_source(&string).map_err(|_| E {
-        msg: "parse error".into(),
-    })?;
-
-    for (fact_str, _) in result.facts {
-        authorizer.add_fact(fact_str)?;
-    }
-    for (rule_str, _) in result.rules {
-        authorizer.add_rule(rule_str)?;
-    }
-    for (check_str, _) in result.checks {
-        authorizer.add_check(check_str)?;
-    }
-
-    for (policy_str, _) in result.policies {
-        authorizer.add_policy(policy_str)?;
-    }
-
-    Ok(())
-}
-
-fn read_private_key_from(from: &KeyBytes) -> Result<PrivateKey, Box<dyn Error>> {
-    let bytes = match from {
-        KeyBytes::FromStdin(KeyFormat::RawBytes) => read_stdin_bytes(),
-        KeyBytes::FromStdin(KeyFormat::HexKey) => {
-            hex::decode(read_stdin_string("hex-encoded private key")?).map_err(|e| e.into())
-        }
-        KeyBytes::FromFile(KeyFormat::RawBytes, path) => fs::read(&path).map_err(|e| e.into()),
-        KeyBytes::FromFile(KeyFormat::HexKey, path) => {
-            hex::decode(fs::read_to_string(&path)?.trim()).map_err(|e| e.into())
-        }
-        KeyBytes::HexString(str) => hex::decode(&str).map_err(|e| e.into()),
-    };
-    PrivateKey::from_bytes(&bytes?).map_err(|e| {
-        E {
-            msg: format!("invalid private key: {}", e),
-        }
-        .into()
-    })
-}
-
-fn read_public_key_from(from: &KeyBytes) -> Result<PublicKey, Box<dyn Error>> {
-    let bytes = match from {
-        KeyBytes::FromStdin(KeyFormat::RawBytes) => read_stdin_bytes(),
-        KeyBytes::FromStdin(KeyFormat::HexKey) => {
-            hex::decode(read_stdin_string("hex-encoded public key")?).map_err(|e| e.into())
-        }
-        KeyBytes::FromFile(KeyFormat::RawBytes, path) => fs::read(&path).map_err(|e| e.into()),
-        KeyBytes::FromFile(KeyFormat::HexKey, path) => {
-            hex::decode(fs::read_to_string(&path)?.trim()).map_err(|e| e.into())
-        }
-        KeyBytes::HexString(str) => hex::decode(&str).map_err(|e| e.into()),
-    };
-    PublicKey::from_bytes(&bytes?).map_err(|e| {
-        E {
-            msg: format!("invalid public key: {}", e),
-        }
-        .into()
-    })
-}
-
-fn read_biscuit_from(from: &BiscuitBytes) -> Result<UnverifiedBiscuit, Box<dyn Error>> {
-    match from {
-        BiscuitBytes::FromStdin(BiscuitFormat::RawBiscuit) => {
-            UnverifiedBiscuit::from(&read_stdin_bytes()?).map_err(|e| e.into())
-        }
-        BiscuitBytes::FromStdin(BiscuitFormat::Base64Biscuit) => {
-            UnverifiedBiscuit::from_base64(&read_stdin_string("base64-encoded biscuit")?)
-                .map_err(|e| e.into())
-        }
-        BiscuitBytes::FromFile(BiscuitFormat::RawBiscuit, path) => {
-            UnverifiedBiscuit::from(&fs::read(&path)?).map_err(|e| e.into())
-        }
-        BiscuitBytes::FromFile(BiscuitFormat::Base64Biscuit, path) => {
-            UnverifiedBiscuit::from_base64(fs::read_to_string(&path)?.trim()).map_err(|e| e.into())
-        }
-        BiscuitBytes::Base64String(str) => {
-            UnverifiedBiscuit::from_base64(&str).map_err(|e| e.into())
-        }
-    }
-}
+use cli::*;
+use errors::*;
+use input::*;
 
 fn handle_command(cmd: &SubCommand) -> Result<(), Box<dyn Error>> {
     match cmd {
